add maze.py
Added FirstOccupy Node, to calculate origin of whole map from the lowest (x,y) that is unoccupied.

This origin is added as a TF frame, its coordinates are in reference to the "map" TF frame.

TODO: Orientation of map currently is dependent on starting orientation of robot. Need to somehow find edges to orientate the x,y to.